Switching to Bootstrap and using common JS functions
Large update to switch the UI to Bootstrap and to use some common
functions to avoid unnecessary redundancy.

Using Font Awesome icons.

+++ inventory/views_settings.py
@@ -32,17 +32,22 @@ from django.shortcuts import render_to_response
 from django.contrib.auth.decorators import login_required, permission_required
 from django.forms import ModelForm
 from django.template import RequestContext
-from django.http import HttpResponseRedirect
+from django.http import HttpResponseRedirect, HttpResponse, HttpResponseBadRequest
 from django.core.urlresolvers import reverse
 from django.conf import settings
+
+from django.template.loader import render_to_string
 
 import models, forms
 
 from allowance import create_archive, import_archive
 
-from django.contrib.auth.models import User
-from settings.models import UserProfile
-from settings.forms import UserForm, UserProfileForm
+from settings.models import User
+from settings.forms import UserForm
+
+from core.views import settings_links, settings_validation
+
+import json
 
 FUNCTIONS = (
 		(u'00', "Captain"),
@@ -56,65 +61,29 @@ FUNCTIONS = (
 @login_required
 def index(request):
     """Displays differents forms to configure Pharmaship."""
-    links = []
-    for application in settings.INSTALLED_APPS:
-        # Do not take in account Django's applications
-        if "django" in application:
-            continue
-        app_data = {}
-        # By default, adding the general settings application to the dict
-        if application == "settings":
-            app_data['name'] = _("Settings")
-            app_data['url'] = "settings"
-        # Looking if the applications has settings that can be imported
-        elif globals().get("{0}.settings.urls".format(application), True):
-            print "application"
-            app_data['name'] = _(application.capitalize())
-            app_data['url'] = "{0}_settings".format(application)
-        else:
-            continue
-
-        # Set the link as active. NEED IMPROVEMENT.
-        if application == "inventory":
-            app_data['active'] = True
-
-        links.append(app_data)
 
     return render_to_response('pharmaship/settings.html', {
-                        'user': (request.user.last_name + " " +request.user.first_name),
-                        'rank': request.user.profile.get_rank(),
+                        'user': request.user,
                         'title':_("Settings"),
-                        'links': links,
-                        'settingsform': forms.SettingsForm(instance=models.Settings.objects.latest('id')).as_table(),
-                        'exportform': forms.ExportForm().as_table(),
-                        'importform': forms.ImportForm().as_table(),
-                        'locationcreateform' : forms.LocationCreateForm().as_table(),
-                        'locationdeleteform' : forms.LocationDeleteForm().as_table(),
+                        'links': settings_links(),
+                        'settingsform': forms.SettingsForm(instance=models.Settings.objects.latest('id')),
+                        'locationcreateform' : forms.LocationCreateForm(),
+                        'locationdeleteform' : forms.LocationDeleteForm(),
                         },
                         context_instance=RequestContext(request))
-
-@permission_required('inventory')
-def validation(request, form, instance=None):
-    """Validates the form in args."""
-    f = form(request.POST, instance=instance)
-    if f.is_valid(): # All validation rules pass
-        f.save()
-    else:
-        # TODO: Error message
-        return False
 
 @permission_required('inventory.settings.can_change')
 def application(request):
     """Validation of the Application form."""
-    if request.method == 'POST':
-        validation(request, forms.SettingsForm, models.Settings.objects.latest('id'))
+    if request.method == 'POST' and request.is_ajax():
+        return settings_validation(request, forms.SettingsForm, models.Settings.objects.latest('id'))
 
-    return HttpResponseRedirect(reverse('inventory_settings')) # Redirect after POST
+    return HttpResponseNotAllowed(['POST',])
 
 @permission_required('inventory')
 def export_data(request):
     """Exports an allowance in a tar.gz archive."""
-    if request.method == 'POST': # If the form has been submitted
+    if request.method == 'POST' and request.is_ajax(): 
         form = forms.ExportForm(request.POST) # A form bound to the POST data
         if form.is_valid(): # All validation rules pass
             # Process the data in form.cleaned_data
@@ -123,59 +92,56 @@ def export_data(request):
         else:
             return HttpResponseRedirect(reverse('inventory_settings'))
     else:
-        return HttpResponseRedirect(reverse('inventory_settings'))
-
-@permission_required('inventory.settings.can_change')
-def import_data(request):
-    """Import data and do some checks."""
-    if request.method == 'POST': # If the form has been submitted
-        form = forms.ImportForm(request.POST, request.FILES) # A form bound to the POST data
-        if form.is_valid(): # All validation rules pass
-            # Process the data in form.cleaned_data
-            import_file = request.FILES['import_file']
-            return render_to_response('pharmaship/import.html', {
-                        'user': (request.user.last_name + " " +request.user.first_name),
-                        'rank': request.user.profile.get_rank(),
-                        'title':_("Allowance Settings"),
-                        'data': import_archive(import_file),
-                        },
-                        context_instance=RequestContext(request))
-        else:
-            return HttpResponseRedirect(reverse('inventory_settings'))
-    else:
-        return HttpResponseRedirect(reverse('inventory_settings'))
+        return HttpResponseNotAllowed(['POST',])
 
 @permission_required('inventory.location.can_add')
 def create_location(request):
     """Creates a new Location for items in Inventory application."""
-    if request.method == 'POST':
+    if request.method == 'POST' and request.is_ajax(): 
         form = forms.LocationCreateForm(request.POST)
         if form.is_valid():
             primary = form.cleaned_data['primary']
             secondary = form.cleaned_data['secondary']
             # Creation of the instance
-            models.Location(primary=primary, secondary=secondary).save()
-            return HttpResponseRedirect(reverse('inventory_settings'))
+            added = models.Location(primary=primary, secondary=secondary)
+            added.save()
+            # Returning the html element added in order to update the list client-side
+            updated_form = render_to_string('html/form.html', {'form': forms.LocationDeleteForm()})
+            data = json.dumps({'success':_('Data updated'), 'form': updated_form})
+            return HttpResponse(data, content_type="application/json")
         else:
-            return HttpResponseRedirect(reverse('inventory_settings'))
+            errors = dict([(k, [unicode(e) for e in v]) for k,v in form.errors.items()])
+            data = json.dumps({'error': _('Something went wrong!'), 'details':errors})
+            return HttpResponseBadRequest(data, content_type = "application/json")
     else:
-        return HttpResponseRedirect(reverse('inventory_settings'))
+        return HttpResponseNotAllowed(['POST',])
+
 
 @permission_required('inventory.location.can_delete')
 def delete_location(request):
     """Deletes selected Location objects for items in Inventory application."""
-    if request.method == 'POST':
+    if request.method == 'POST' and request.is_ajax():
         form = forms.LocationDeleteForm(request.POST)
         if form.is_valid():
             obj = form.cleaned_data['to_delete']
-
             # Changing values for objects with location in obj
             default = models.Location.objects.get(pk=1)
             models.Medicine.objects.filter(location__in=obj).update(location = default)
             models.Article.objects.filter(location__in=obj).update(location = default)
+            # Get the list of ids to be deleted in order to update the list client-side
+            obj_id = []
+            for item in obj:
+                obj_id.append(item.pk)
+            # Delete
             obj.delete()
-            return HttpResponseRedirect(reverse('inventory_settings'))
+            # Returning the updated list of locations
+            print "OBJ", obj_id
+
+            data = json.dumps({'success':_('Data updated'), 'deleted': obj_id})
+            return HttpResponse(data, content_type="application/json")
         else:
-            return HttpResponseRedirect(reverse('inventory_settings'))
+            errors = dict([(k, [unicode(e) for e in v]) for k,v in form.errors.items()])
+            data = json.dumps({'error': _('Something went wrong!'), 'details':errors})
+            return HttpResponseBadRequest(data, content_type = "application/json")
     else:
-        return HttpResponseRedirect(reverse('inventory_settings'))
+        return HttpResponseNotAllowed(['POST',])
